requirements.txt
autopep8==1.5.7
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
comtypes==1.1.10
idna==2.10
MouseInfo==0.1.3
Pillow==8.2.0
PyAudio @ file:///C://Users//cristhian.torres//Documents//projects//python/PyAudio-0.2.11-cp39-cp39-win_amd64.whl
PyAutoGUI==0.9.52
pycodestyle==2.7.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
pypiwin32==223
PyRect==0.1.4
PyScreeze==0.1.27
python-decouple==3.4
pyttsx3==2.90
PyTweening==1.0.3
pywhatkit==4.8
pywin32==301
requests==2.25.1
soupsieve==2.2.1
SpeechRecognition==3.8.1
toml==0.10.2
urllib3==1.26.5
wikipedia==1.4.0